Changed the data files to more standard formats.
The header of a TSV file should generally be accesible and not commented out. I
removed the comment on the header line and added tabs to delimit the headers. I
then switch the load call to importdata which deals with the headers in the
correct way. Finally, I removed the addpath call which is not best practice.